pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "plato-cli"
version = "0.1.0"
description = "PLATO CLI — command-line interface for all consciousness packages"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28",
    "plato-room-phi>=0.1.0",
    "plato-attention-tracker>=0.1.0",
    "plato-meta-tiles>=0.1.0",
    "plato-surrogate>=0.1.0",
    "plato-fflearning>=0.1.0",
    "plato-surprise-detector>=0.1.0",
    "seed-creative-swarm>=0.1.0",
    "flux-reasoner>=0.1.0",
    "fleet-consciousness-dashboard>=0.1.0",
]

[project.scripts]
plato-cli = "plato_cli:main"